requirements.txt

streamlit
pytesseract
Pillow
